requirements.txt
torch
transformers
datasets
peft
accelerate
trl
bitsandbytes
huggingface_hub
litellm
openai
fastapi
pydantic>=2.0.0
pyyaml
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
langgraph
requests
firecrawl-py
